Fix media-view:updateURLS() to handle url not set edge case.
There is an edge-case where currentSrc isn't set yet, but everything
else about the video is valid. The workaround is same as that done in
core/media.js, which is to we wait for it to be set before calling
the updateURLS function.
Without this, media.url is undefined and setupMedia throws an error.